+++ src/helpers.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * This file is part of Pipe package.
  *
@@ -8,14 +9,16 @@
 declare(strict_types=1);
 
 use Serafim\Pipe\Pipe;
+use Serafim\Pipe\PipeInterface;
 
 if (! \function_exists('pipe')) {
     /**
      * @param mixed|null $value
-     * @return Pipe
+     * @param string|null $namespace
+     * @return Pipe|PipeInterface
      */
-    function pipe($value = null): Pipe
+    function pipe($value = null, string $namespace = null): PipeInterface
     {
-        return new Pipe($value);
+        return new Pipe($value, $namespace);
     }
 }